First attempt at using a SparseGrid.
The idea is to use SLSRow as the RecycleView widget. SLSRow is a subclass of
SparseGridLayout and has a list property that holds the widgets. When the class
is instantiated, these widgets are added to the SparseGridLayout. The SLSRow
grid is hard-coded to (3,1). Hopefully, at a later point, we can use this to
make the app responsive.